[Slider]: moved in uui and made PE for all themes
