+++ app/src/docs/Slider.doc.tsx
@@ -1,9 +1,5 @@
 import * as React from 'react';
-import {
-    BaseDocsBlock, DocExample, EditableDocContent, TDocsGenType, UUI3,
-} from '../common';
-import cx from 'classnames';
-import css from './styles.module.scss';
+import { BaseDocsBlock, DocExample, EditableDocContent, TDocsGenType, UUI3, UUI4, UUI } from '../common';
 
 export class SliderDoc extends BaseDocsBlock {
     title = 'Slider';
@@ -13,16 +9,18 @@ export class SliderDoc extends BaseDocsBlock {
     getPropsDocPath() {
         return {
             [UUI3]: './app/src/docs/_props/loveship/components/inputs/Slider/slider.props.ts',
+            [UUI4]: './app/src/docs/_props/epam-promo/components/inputs/slider.props.ts',
+            [UUI]: './app/src/docs/_props/uui/components/inputs/slider.props.ts',
         };
     }
 
     renderContent() {
         return (
-            <span className={ cx(css.wrapper, css.uuiThemePromo) }>
+            <>
                 <EditableDocContent fileName="slider-descriptions" />
                 {this.renderSectionTitle('Examples')}
                 <DocExample title="Basic" path="./_examples/slider/Basic.example.tsx" />
-            </span>
+            </>
         );
     }
 }
