+++ app/src/docs/_examples/mainMenu/Responsive.example.tsx
@@ -1,6 +1,5 @@
 import React, { ReactNode, useState } from 'react';
-import { BurgerButton, GlobalMenu, MainMenu, MainMenuAvatar, MainMenuButton, MainMenuIcon, FlexSpacer, FlexCell, MainMenuDropdown, BurgerSearch, DropdownMenuBody, DropdownMenuButton, DropdownMenuSplitter, Burger } from '@epam/uui';
-import { Slider } from '@epam/promo';
+import { Slider, BurgerButton, GlobalMenu, MainMenu, MainMenuAvatar, MainMenuButton, MainMenuIcon, FlexSpacer, FlexCell, MainMenuDropdown, BurgerSearch, DropdownMenuBody, DropdownMenuButton, DropdownMenuSplitter, Burger } from '@epam/uui';
 import { Dropdown, MainMenuLogo, AdaptiveItemProps } from '@epam/uui-components';
 import { ReactComponent as HelpIcon } from '@epam/assets/icons/common/notification-help-outline-24.svg';
 import { ReactComponent as PinIcon } from '@epam/assets/icons/common/action-pin_on-24.svg';
