+++ app/src/docs/_examples/slider/Basic.example.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
-import { FlexCell } from '@epam/uui';
-import { Slider } from '@epam/loveship';
+import { FlexCell, Slider } from '@epam/uui';
 import css from './BasicExample.module.scss';
 
 export default function BasicExample() {
