+++ app/src/docs/_props/uui/components/inputs/slider.props.ts
@@ -1,7 +1,7 @@
 import { DocBuilder } from '@epam/uui-docs';
 import { SliderBaseProps } from '@epam/uui-components';
-import { Slider } from '@epam/loveship';
-import { DefaultContext, FormContext, ResizableContext, iEditable, isDisabledDoc } from '../../../docs';
+import { Slider } from '@epam/uui';
+import { DefaultContext, ResizableContext, iEditable, isDisabledDoc } from '../../docs';
 
 const sliderDoc = new DocBuilder<SliderBaseProps<number>>({ name: 'Slider', component: Slider })
     .implements([
@@ -49,6 +49,6 @@ const sliderDoc = new DocBuilder<SliderBaseProps<number>>({ name: 'Slider', comp
         ],
         isRequired: false,
     })
-    .withContexts(DefaultContext, ResizableContext, FormContext);
+    .withContexts(DefaultContext, ResizableContext);
 
 export default sliderDoc;
