--- epam-promo/components/inputs/Slider/index.ts
@@ -1,1 +1,0 @@
-export * from './Slider';
