+++ epam-promo/components/inputs/index.tsx
@@ -1,4 +1,3 @@
 export * from './MultiSwitch';
 export * from './Rating';
 export * from './TextInput';
-export * from './Slider';
