--- loveship/components/inputs/Slider/Slider.tsx
@@ -1,9 +1,0 @@
-import { Slider as UuiSlider, SliderBaseProps } from '@epam/uui-components';
-import css from './Slider.module.scss';
-import { withMods } from '@epam/uui-core';
-
-export function applySliderMods() {
-    return [css.root];
-}
-
-export const Slider = withMods<SliderBaseProps<number>>(UuiSlider, applySliderMods);
