+++ loveship/components/inputs/Slider/index.ts
@@ -1,2 +1,1 @@
 export * from './RangeSlider';
-export * from './Slider';
