+++ uui/components/inputs/Slider.module.scss
@@ -1,5 +1,3 @@
-// Not used in projects
-
 .root {
     :global(.uui-tooltip-body) {
         background-color: var(--uui-neutral-70);
