+++ uui/components/inputs/__tests__/Slider.test.tsx
