+++ uui/components/inputs/index.ts
@@ -7,3 +7,4 @@ export * from './NumericInput';
 export * from './TextArea';
 export * from './TimePicker';
 export * from './InputAddon';
+export * from './Slider';
